Improve performance in IncrementalItemCF::recommend().
For the candidate items, the previous implementation includes all
items which has similarity value with input items.

Now they are limited to top neighbors with input items, it could
reduce large candidates, especially when similarity matrix is dense.

From the test on movielen data, it shows that the TPS is improved from 59 to 786.