make IncrementalItemCF functions thread safe.
in previous implementation, it calls like below to read a row:
1. boost::shared_ptr<const VisitRow> coVisitRow = visitMatrix_.row(row);
2. iterate on above "coVisitRow"

as another thread might be calling "visitMatrix_.update_row_with_func()",
the 2nd step is not thread safe, because it might be reading a row which
is being modified.

now all of the row reading operations are revised to call
"MatrixDB::read_row_with_func()".